% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gene_report.R
\name{gene_report_amplification}
\alias{gene_report_amplification}
\title{Generates a gene amplification report using a list of CNV files located in cnv_base_path.}
\usage{
gene_report_amplification(cnv_base_path, num_genes_top)
}
\arguments{
\item{cnv_base_path}{A path to all relevant CNV files for analysis.}

\item{num_genes_top}{Integer which specifies # of highest-frequency genes to return.}
}
\value{
Frequency table of amplified genes, with num_genes_top rows
}
\description{
Generates a gene amplification report using a list of CNV files located in cnv_base_path.
}
