% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gene_report.R
\name{get_amplification}
\alias{get_amplification}
\title{Returns binded data.frame of all amplified genes in a given CNV file.}
\usage{
get_amplification(patient_file)
}
\arguments{
\item{patient_file}{A data.frame containing single patient data from CNV file.}
}
\value{
The total data.frame of all amplified genes for given patient
}
\description{
Returns binded data.frame of all amplified genes in a given CNV file.
}
